Move UX code to its own module
